requirements.txt
pandas>=2.2
pyarrow>=16.0
yfinance>=0.2.52
fredapi>=0.5.2
python-dotenv>=1.0.1
pyyaml>=6.0.2
